requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
boto3==1.34.138
pandas==2.2.2
psycopg2==2.9.9
requests==2.32.3
openpyxl==3.1.5
pydantic-settings==2.4.0
langchain==0.0.349
langchain_core==0.0.13
langchain_community==0.0.1
faiss-cpu==1.8.0.post1
SQLAlchemy==2.0.32
PyMySQL==1.1.1
